Some explicit casts after removal of "long" overloads from QString::arg and QByteArray::arg.
[git-p4: depot-paths = "//depot/qt/main/": change = 165676]

+++ extensions/activeqt/container/qaxdump.cpp
@@ -42,7 +42,7 @@ static QString docuFromName(ITypeInfo *typeInfo, const QString &name)
             if (!docString.isEmpty())
                 docu += docString + "\n";
             if (!helpFile.isEmpty())
-                docu += QString("For more information, see help context %1 in %2.\n").arg(helpContext).arg(helpFile);
+                docu += QString("For more information, see help context %1 in %2.\n").arg((uint)helpContext).arg(helpFile);
         }
     }
